input images in order
These images are taken by me inside
the college campus with prior notice
to the management and with the
consent of the respective vehicle
owners.
Although only entry time of the
vehicles could be captured, for
better simulating vehicle movement
and for the purpose to keep a track
of vehicle occupancy in the parking
space, exit times of the vehicles
are modified according to the needs.